Some BMPs reduce runoff from total area runoff
Introduces a change to allow a subset of BMPs to reduce runoff "off the
top".  Because water quality calculations are based on runoff at the
leaf level (individual soil:LU pairings) I actually reduced the runoff
at each pairing level so that the sums of both runoff|inf|et and water
quality are equally reduced over the whole area.  The top level is
reduced implicitly.

The infiltration or runoff factor is yet to be developed, so the
implementation of how the reduction is applied may need to change again
in the future, but this sets up the infrastructure to accommodate that
change regardless.